requirements.txt
biopython
scikit-learn
pandas
matplotlib
joblib
numpy

